Outliner: Port scene elements and some related types to new tree-element code design
Continuation of work in 2e221de4ceee, 249e4df110e0 and 3a907e742507.

Adds new tree-element classes for the scene-ID, scene collections, scene
objects, and the view layers base.
There is some more temporary stuff in here, which can be removed once we're
further along with the porting. Noted that in comments.